🧐 [Scraper] added another string comparison method
Copilot suggested the JaroWinkler algo, and this is it.
See SimilarityTests for a comparison output.
Routed all other direct calls of the old method into MetadataUtil.
Old method still called, but max`ed with new one, getting even better
result:

So a "Avatar" and "Avatar (2009)" comparison is now a nearly 90% match,
where in our old impl we wouldn't have taken that result because of <75%
score... (66%, in fact)